requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.35.0
python-dotenv>=1.0.0
httpx>=0.27.0
groq>=0.9.0
google-genai>=1.0.0
mistralai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
networkx>=3.3
chromadb>=0.5.0
sentence-transformers>=3.0.0
tree-sitter>=0.22.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-java>=0.21.0
PyGithub>=2.3.0
gitpython>=3.1.0
jinja2>=3.1.0
pydantic>=2.7.0
